Update 1. Create Staging Internetsales  Table.sql

+++ DW/1. Create Staging Internetsales  Table.sql
@@ -4,15 +4,15 @@ GO
 
 CREATE TABLE staging.internetsales
 (
-	ProductKey bigint  NULL,
-	OrderDateKey bigint  NULL,
-	CustomerKey bigint  NULL,
-	OrderQuantity bigint  NULL,
-	UnitPrice float  NULL,
-	UnitPriceDiscountPct bigint  NULL,
-	DiscountAmount bigint  NULL,
-	TotalProductCost float  NULL,
-	SalesAmount float  NULL,
-	TaxAmt float  NULL
+	ProductKey int  NULL,
+	OrderDateKey int  NULL,
+	CustomerKey int  NULL,
+	OrderQuantity int  NULL,
+	UnitPrice decimal(18,2)  NULL,
+	UnitPriceDiscountPct decimal(8,2)  NULL,
+	DiscountAmount decimal(18,2)  NULL,
+	TotalProductCost decimal(18,2)  NULL,
+	SalesAmount decimal(18,2)  NULL,
+	TaxAmt decimal(18,2)  NULL
 )
 GO
